pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlassian-cli"
version = "0.1.0"
description = "CLI tool to synchronize Confluence spaces with local folders"
readme = "README.md"
authors = [{name = "Your Name", email = "you@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-frontmatter>=1.0.0",
    "markdown-it-py>=3.0.0",
    "markdownify>=0.12.0",
    "rich>=13.7.0",
]

[project.scripts]
atlassian-cli = "atlassian_cli.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
extend-select = ["I"]

